dt-bindings: clock: Added documentation for global STM time source
The documentation file explains that the particular usecase for the STM is
as a global time source for the entire SoC. A compatible device should
contain the base address for this peripheral, alongside reference to the
name of the clock groups (stm).

Issue: ALB-5093 ALB-8823 ALB-8965

Signed-off-by: Dorin Ionita <dorin.ionita@nxp.com>
Signed-off-by: Andrei Botila <andrei.botila@nxp.com>
Signed-off-by: Radu Pirea <radu-nicolae.pirea@nxp.com>

devicetree/bindings: Document STM compatible

The compatible string "nxp,s32cc-stm-ts" will be used by the
STM instances that assume the counting will be started on the
driver's probe.

Issue: ALB-10872
Issue: ALB-9261
Signed-off-by: Ghennadi Procopciuc <ghennadi.procopciuc@nxp.com>