s32cc: nvmem_common: Refactor `read_nvmem_cell()` helper
Simplified the `read_nvmem_cell()` helper such that it does
not require the caller to use a char* to get the return value,
like the original `nvmem_cell_read()` call.

The whole point of the helper is to make life easier for the
caller, and all s32cc-related NVMEM cells use an `u32` size,
thus, we can directly set a passed `u32*` inside the helper
instead of returning a pointer to a buffer and letting the caller
extract its `u32` value from there.

Issue: ALB-10366
Signed-off-by: Andrei Cherechesu <andrei.cherechesu@nxp.com>