arm_scmi: smc: save flags only if spin lock was acquired
Previously, if we had concurrent requests, we could end up in a
situation were incorrect flags were restored which resulted in a
core running with interrupts disabled. Then, schedutil would try to
run and discover that interrupts were disabled and it would, in turn,
issue the bug message.

Issue: ALB-10746
Issue: ALB-10894

Signed-off-by: Ghennadi Procopciuc <ghennadi.procopciuc@nxp.com>
Signed-off-by: Andrei Stefanescu <andrei.stefanescu@nxp.com>