dts: s32g: add pinctrl settings for UART1
Add the necessary pinctrl settings for UART1 and enable it by default
for rdb boards. The evb boards share one pin with the I2C so only added
the pinctrl settings but left the UART1 disabled.

The serial to be used is filled by U-Boot via the stdout-path in
the /chosen node.

Issue: ALB-10945
Signed-off-by: Andrei Stefanescu <andrei.stefanescu@nxp.com>