iio: adc: s32cc_adc: fix syncing of dma region
When using CONFIG_DMA_API_DEBUG, the kernel throws warning about the
ownership of DMA memory.
This happens because the syncing of the dma region is done by the ADC
driver. This is incorect because that region is allocated to the dma
driver.

Use the dma device instead of the ADC device.

Issue: ALB-11315
Signed-off-by: Alexandru-Catalin Ionita <alexandru-catalin.ionita@nxp.com>