iio: adc: s32cc_adc: Optimize DMA callback
During a stress test of the ADC it was observed a stall.

If the thread who disables the ADC does not get the chance to execute,
since it has interrupts disabled, it will trigger an RCU stall.

The disabling thread does not get a chance to execute because the
DMA's virtual channel's spin lock is hold by the DMA's descriptor
callback which execute s32cc_adc_dma_cb.

There is an off by one error in the logic of the ADC DMA callback.
If the head of the circular buffer gets to be the size of
buffer - 4096 - the while loop never finishes.

Correct this error.

Issue: ALB-11492
Signed-off-by: Alexandru-Catalin Ionita <alexandru-catalin.ionita@nxp.com>