gpio: siul2: populate num_reg_defaults_raw for output regmaps
Previously, the cache for the PGPDO registers would be populated with 0
because no reg_defaults/reg_defaults_raw were given to the
regmap_config.

Force the regcache to read the hardware state by passing in
num_reg_defaults_raw set to the number of registers for that regmap.

Issue: ALB-11725

Signed-off-by: Andrei Stefanescu <andrei.stefanescu@nxp.com>