s32cc: serdes: Fix SerDes clock check on init.
The SerDes RX clock check on init had two problems. First, it
requires non-zero time after XPCS init for clocks to lock, i.e., it
reported unstable RX when nothing was unstable.
Second, it was called in the XPCS reset function. At the time of XPCS
reset it is not clear which XPCS is supposed to do what, i.e., it is
not visible if an XPCS that could be used is in fact disabled.
Given that the reset code resets both XPCS even though on G2 in mode 4
only 1 may be used, you got invalid error messages on RX clock check.

The solution was to factor out a function that does that check with
a proper timeout loop and to only check when the SerDes PHY init is
called for lanes that are actually used.

Even though this change means that due to UART output the function will
likely report stable RX after 0µs, the actual time after PLL reset is
multiple microseconds. Given that UART messages may be suppressed, the
timeout loop cannot be eliminated even though it may seem 0µs at first
glance.

Issue: ALB-11741
Signed-off-by: Heinz Wrobel <Heinz.Wrobel@nxp.com>
Signed-off-by: Ciprian Costea <ciprianmarian.costea@nxp.com>