arch: fdts: VC-BASE: add support for SFP cage
SFP cage is over I2C3. Add pinmux for I2C3.
Even though, when I2C3 is enabled, the SFP is detected and it's EEPROM
is read, the SFP is detected as a fiber tranciever - (2500baseX).
Since our PCS does not support 2500base-X autoneg, we disable the I2C3
and use a fixed-link at 2500 speed.

Issue: ALB-11935
Signed-off-by: Alexandru-Catalin Ionita <alexandru-catalin.ionita@nxp.com>